feat: general improvements to the compiler

+++ src/compiler/files/image.js
@@ -11,15 +11,11 @@ const imageminPngquant = require('imagemin-pngquant')
 module.exports = class Image extends File {
 
     constructor(filename, project) {
-
         super(filename, project)
     }
 
     optimize(destination) {
-
         return imagemin([this.filename], destination, {
-
-
             plugins: [
                 imageminJpegtran({
                     progressive: true
@@ -32,56 +28,46 @@ module.exports = class Image extends File {
     }
 
     rename(oldPath, newPath) {
-
         return new Promise((resolve, reject) => {
-
             fs.rename(oldPath, newPath, e => {
-
                 if (e) return reject(e)
-
                 resolve()
             })
         })
     }
 
     rmdir(destination) {
-
         return new Promise((resolve, reject) => {
-
             fs.rmdir(destination, e => {
-
                 if (e) return reject(e)
-
                 resolve()
             })
         })
     }
 
-    async render() {
+    async render($optimize = true) {
+        if ($optimize === true) {
+            const file = this.filename.replace(this.project.path, 'build')
+            const paths = path.parse(file)
+            const target = `${paths.dir}/opt-${paths.name}${this.exportExtension}`
+            const destination = `${this.project.path}/${target}`
+            const files = await this.optimize(destination)
 
-        const file = this.filename.replace(this.project.path, 'build')
-        const paths = path.parse(file)
-        const target = `${paths.dir}/opt-${paths.name}${this.exportExtension}`
-        const destination = `${this.project.path}/${target}`
-
-        const files = await this.optimize(destination)
-
-        if (files[0]) {
-
-            try {
-                const newPath = files[0].path.replace(`opt-${paths.name}${this.exportExtension}/`, '')
-                await this.rename(files[0].path, newPath)
-                await this.rmdir(destination)
-
-            } catch (err) {
-
-                return Promise.reject(err.message, this.filename)
+            if (files[0]) {
+                try {
+                    const newPath = files[0].path.replace(`opt-${paths.name}${this.exportExtension}/`, '')
+                    await this.rename(files[0].path, newPath)
+                    await this.rmdir(destination)
+                } catch (err) {
+                    return Promise.reject(err.message, this.filename)
+                }
             }
+        } else {
+            this.copy()
         }
     }
 
     get exportExtension() {
-
         // Keep the same extension
         return path.parse(this.filename).ext
     }
